+++ src/BotService/Changelog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed issue where New-AzBotService fails to convert ApplicationSecret to plain text on Windows PowerShell [#16131]
 
 ## Version 0.3.1
 * Fixed the issue that `New-AzBotService` will fail in WebApp mode. [#15608]